Protect some /posts routes with authentication
- Create
- Update
- Delete

+++ server/src/routes/posts/delete.js
@@ -6,12 +6,19 @@ const deletePost = async (req, res) => {
     if (!mongoose.Types.ObjectId.isValid(req.params.id)) {
       return res.status(400).json();
     }
-    // TODO: In the future, prevent some fields from being updated
-    const result = await postModel.deleteOne({_id: req.params.id});
-    if (result.deletedCount && result.acknowledged) {
-      return res.status(200).json(result);
+
+    const post = await postModel.findById(req.params.id);
+
+    if (!post) {
+      return res.status(404).send();
     }
-    return res.status(404).json(result);
+
+    if (post.userId != req.user.userId) {
+      return res.status(403).json({error: 'Forbidden'});
+    }
+
+    const result = await postModel.deleteOne({_id: req.params.id});
+    return res.status(200).json(result);
   } catch (error) {
     console.log(error);
     return res.status(500).send();
